fix: reduce initial bundle size and improve load order
1. Remove loading graphic as it stalls js loading until complete
2. Replace heavyweight CID/PeerId parsing libraries with dependency-free Cloudflare snippet version
3. Write final js/css filenames into source code to eliminate extra round-trips to read build config
4. Refactor Cloudflare snippets to be build from ts source and tested
5. Refactor Cloudflare helpers to make them importable from rest of source

Removing the image saves 45KB, using the Cloudflare snippet utilities
reduces the best-case bundle (e.g. a valid CID needing a redirect)
from 250KB to about 10KB.

Using esbuild to build the Cloudflare snippets also reduces their size,
for example 01 goes from 9KB down to about 3KB.
