+++ src/cloudflare/snippets/02_shared_sw_installer_cache.js.rules
