+++ src/cloudflare/snippets/dnslink.ts
@@ -1,0 +1,8 @@
+/**
+ * DNSLink label encoding
+ *
+ * @see https://specs.ipfs.tech/http-gateways/subdomain-gateway/#host-request-header
+ */
+export function dnsLinkEncode (domain: string): string {
+  return domain.replace(/-/g, '--').replace(/\./g, '-')
+}
